updated Desktop gui: now only shows one spin if only one present
